backend/requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
asyncpg
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
bcrypt
psycopg2-binary

